Adding coordinates displacer
Adding functionality to displace a selected power station assets away from each. This allows all the assets for a particular power station to be visible on the map. Without this functionality the assets become stacked on the map and only details on the asset at the top of the stack can be seen since they all share the same coordinates.